bashrc: Make ls invocation BSD/Solaris compatible
On BSD ls -Ahs doesn't print a total, because -h isn't considered
unless -l is there too. So use ls -Alhs instead.

FreeBSD prints the total number of *blocks* though:

    eschaton git (next) $ ls -lha | grep total
    total 354522
    eschaton git (next) $ ls -As | awk '// {total = total + $1} END {print total}'
    354506

So it's off by a bit, but still gives an idea of the size.

This also works on Solaris /bin/ls, which appears to print bytes, not
blocks.