.bashrc: nuke ~/.gitconfig if the only change to ~/ is an E-Mail change there
I have something that changes my E-Mail in ~/.gitconfig to my work
E-Mail address, but this makes updating annoying. Change the updater
to nuke ~/.gitconfig if the only change is an E-Mail change, then
it'll be reset at the end of dud when we run the bashrc.

This would be much better with the config inclusion support, but we
haven't upgraded to that version of git yet.