chansort_configurable: Fix a bug & change the default
For the reasons covered in the commets I'm adding here the
${a,b}.{type,chatnet,name} would sometimes not be filled, which my
default sorter was assuming so it would end up spewing some warnings.

So change the filling of the defaults to that this is always filled, and
document what the value passed as $a or $b will look like.

Finally, change the default sorting from the more naïve example that
just mostly does alphabetic sorting to one that preserves the order of
the QUERY windows as they're created.