offlineimap: stop using the tunnel
The immediate reason for this is that I can't start it anymore, I get:

    [!] /home/avar/.stunnel.work-gmail-imap:10:
    "[work-imap-gmail-ssl-tunnel]": Failed to initialize TLS

Too lazy to debug this now, but while I'm at it just disable it. Like
1/4 of my home network's traffic goes to IMAP because of this tunnel
disabling compression, and I can always get this working in the future
if I'd like to debug the IMAP traffic again.