.gitconfgig: (broken) test setup with hyperfine
Make it easy to test HEAD~ v.s. HEAD in rebase --exec, e.g.:

     git rebase -i --exec 'git hf-cmp-prev-all-all-warmup --show-output'

There's a bug here with the "git checkout" that I'm about to fix.  But
let's commit the pre-image, somehow running the "checkout" in the
worktree yields conflicts on a local rebase, hrm?

This relies on my https://github.com/sharkdp/hyperfine/pull/447
https://github.com/sharkdp/hyperfine/pull/448