local screen: add public-inbox git-ml indexing
Follow-up my 739ec1c (add public-inbox to PATH, 2022-02-01). This is
an attempt to follow the instructions at
https://lore.kernel.org/git/20211206163626.GA4714@dcvr/

The initial indexing took a long time (hour?) and created:

    $ pwd; du -sh .git/public-inbox/
    /home/avar/g/git-ml
    5.5G    .git/public-inbox/

But now:

    $ public-inbox-index -v $PWD
    # indexing /home/avar/g/git-ml/.git ...
    # counting changes
            fe3d014914cecf8eab486e20d4b7ef686c292050..HEAD ... # 0

So nothing to do, but then after a git pull:

    $ public-inbox-index -v $PWD
    # indexing /home/avar/g/git-ml/.git ...
    # counting changes
            fe3d014914cecf8eab486e20d4b7ef686c292050..HEAD ... # 2
    # indexed 2/2

Which also took almost no time at all.

The "stuff" commands show the rest of the commands needed to get this
up and running.